[ng] date moved adapter to /core
