+++ packages/ng/libraries/core/src/lib/date/index.ts
@@ -1,3 +1,4 @@
 export * from './date-granularity.enum';
 export * from './date-adapter.interface';
 export * from './date-adapter.class';
+export * from './native/index';
