+++ packages/ng/libraries/core/src/lib/date/native/index.ts
@@ -1,0 +1,2 @@
+export * from './native-date.option';
+export * from './native-date.adapter';
