+++ packages/ng/libraries/date/src/lib/index.ts
@@ -3,6 +3,4 @@ export * from './calendar/index';
 export * from './input/index';
 export * from './picker/index';
 export * from './adapter/index';
-export * from './native/index';
-export * from './native-utc/index';
 export * from './select/index';
