--- packages/ng/libraries/date/src/lib/native-utc/index.ts
@@ -1,2 +1,0 @@
-export * from './native-utc-date.adapter';
-export * from './native-utc-date.module';
