--- packages/ng/libraries/date/src/lib/native-utc/native-utc-date.module.ts
@@ -1,11 +1,0 @@
-import { NgModule } from '@angular/core';
-import { ALuDateAdapter } from '@lucca-front/ng/core';
-import { LuNativeUTCDateAdapter } from './native-utc-date.adapter';
-
-@NgModule({
-	providers: [
-		LuNativeUTCDateAdapter,
-		{ provide: ALuDateAdapter, useClass: LuNativeUTCDateAdapter },
-	]
-})
-export class LuNativeUTCDateModule {}
