--- packages/ng/libraries/date/src/lib/native/index.ts
@@ -1,2 +1,0 @@
-export * from './native-date.adapter';
-export * from './native-date.module';
